:art: Move SQL functions from highscores.js to models.js
Rename user.js (from models) into model.js (moved into libs)
